Replacing Agent views and menu items with Agent Modal: - reorganizing Agent modal, AgentCard, and related files - replacing old routes to list and editor with redirects to chain editor - removing all unused agent files - adding Modal close context so buttons in the modal can close it - adding Agent modal to ChainEditorView menu
