--- frontend/agents/AgentMetricsPanel.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const AgentMetricsPanel = ({ agent }) => {
-  return null;
-};
