--- frontend/agents/AgentsListView.js
@@ -1,28 +1,0 @@
-import React from "react";
-import { Layout, LayoutContent } from "site/Layout";
-import { AgentCardList } from "agents/AgentCardList";
-import { Heading, Spinner, VStack } from "@chakra-ui/react";
-import { ScrollableBox } from "site/ScrollableBox";
-import { usePaginatedAPI } from "utils/hooks/usePaginatedAPI";
-import { useLocation } from "react-router-dom";
-
-export const AgentsListView = () => {
-  const location = useLocation();
-  const { page, isLoading } = usePaginatedAPI("/api/agents/", {
-    loadDependencies: [location],
-    limit: 90000,
-  });
-
-  return (
-    <Layout>
-      <LayoutContent>
-        <VStack alignItems="start" p={5} spacing={4}>
-          <Heading>Agents</Heading>
-          <ScrollableBox>
-            {isLoading ? <Spinner /> : <AgentCardList page={page} />}
-          </ScrollableBox>
-        </VStack>
-      </LayoutContent>
-    </Layout>
-  );
-};
