+++ frontend/chains/editor/sidebar/EditorAgentCard.js
@@ -1,14 +1,15 @@
 import React from "react";
 
 import { AgentCard } from "agents/AgentCard";
-import { AgentToggleButton } from "chat/agents/AgentToggleButton";
+import { ModalClose } from "components/Modal";
 import { AgentEditButton } from "agents/AgentEditButton";
 
-export const ChatAgentCard = ({ agent, ...props }) => {
+export const EditorAgentCard = ({ agent, ...props }) => {
+  const close = React.useContext(ModalClose);
+
   return (
     <AgentCard agent={agent} {...props}>
-      <AgentEditButton agent={agent} />
-      <AgentToggleButton agent={agent} />
+      <AgentEditButton agent={agent} onClick={close} />
     </AgentCard>
   );
 };
